Combine ExternalUser and ExternalTeam model serializers
